move ComputePackageSCCs to a new file
